Refactory create appointment
